requirements.txt
streamlit==1.29.0
tensorflow==2.15.0
requests==2.28.1
Pillow==8.3.2